+++ packages/svelte-router/src/not-found/DefaultGlobalNotFound.svelte
@@ -1,0 +1,1 @@
+<p>Not Found</p>
